refctor: remove old sql
