+++ src/service/search/cache/cacher.rs
@@ -32,8 +32,7 @@ use crate::{
             result_utils::{get_ts_value, round_down_to_nearest_minute},
             MultiCachedQueryResponse,
         },
-        new_sql::NewSql,
-        sql::{generate_histogram_interval, RE_HISTOGRAM, RE_SELECT_FROM},
+        new_sql::{generate_histogram_interval, NewSql, RE_HISTOGRAM, RE_SELECT_FROM},
     },
 };
