+++ src/service/search/grpc/storage.rs
@@ -22,7 +22,7 @@ use config::{
         bitvec::BitVec,
         inverted_index::search::{ExactSearch, PrefixSearch, SubstringSearch},
         search::{ScanStats, StorageType},
-        stream::{FileKey, PartitionTimeLevel, StreamPartition, StreamType},
+        stream::FileKey,
     },
     utils::inverted_index::{
         convert_parquet_idx_file_name, create_index_reader_from_puffin_bytes, split_token,
@@ -45,10 +45,7 @@ use tracing::Instrument;
 
 use crate::service::{
     db, file_list,
-    search::{
-       datafusion::exec, generate_filter_from_equal_items,
-        generate_search_schema_diff, sql::Sql,
-    },
+    search::{datafusion::exec, generate_filter_from_equal_items, generate_search_schema_diff},
 };
 
 type CachedFiles = (usize, usize);
@@ -296,54 +293,6 @@ pub async fn search(
     }
 
     Ok((tables, scan_stats))
-}
-
-#[tracing::instrument(name = "service:search:grpc:storage:get_file_list", skip_all, fields(org_id = sql.org_id, stream_name = sql.stream_name))]
-async fn get_file_list(
-    trace_id: &str,
-    sql: &Sql,
-    stream_type: StreamType,
-    time_level: PartitionTimeLevel,
-    partition_keys: &[StreamPartition],
-) -> Result<Vec<FileKey>, Error> {
-    log::debug!(
-        "[trace_id {trace_id}] search->storage: get file_list in grpc, stream {}/{}/{}, time_range {:?}",
-        &sql.org_id,
-        &stream_type,
-        &sql.stream_name,
-        &sql.meta.time_range
-    );
-    let (time_min, time_max) = sql.meta.time_range.unwrap();
-    let file_list = match file_list::query(
-        &sql.org_id,
-        &sql.stream_name,
-        stream_type,
-        time_level,
-        time_min,
-        time_max,
-    )
-    .await
-    {
-        Ok(file_list) => file_list,
-        Err(err) => {
-            log::error!("[trace_id {trace_id}] get file list error: {}", err);
-            return Err(Error::ErrorCode(ErrorCodes::ServerInternalError(
-                "get file list error".to_string(),
-            )));
-        }
-    };
-
-    let mut files = Vec::with_capacity(file_list.len());
-    for file in file_list {
-        if sql
-            .match_source(&file, false, false, stream_type, partition_keys)
-            .await
-        {
-            files.push(file.to_owned());
-        }
-    }
-    files.sort_by(|a, b| a.key.cmp(&b.key));
-    Ok(files)
 }
 
 #[tracing::instrument(name = "service:search:grpc:storage:cache_files", skip_all)]
