+++ src/service/search/super_cluster/leader.rs
@@ -233,7 +233,7 @@ async fn run_datafusion(
         let table_name = sql.stream_names.first().unwrap();
         physical_plan = Arc::new(RemoteScanExec::new(
             physical_plan,
-            rewrite.file_lists.get(table_name).unwrap().clone(),
+            rewrite.file_id_lists.get(table_name).unwrap().clone(),
             rewrite.idx_file_list.clone(),
             rewrite
                 .equal_keys
